+++ spade-tests/src/integration.rs
@@ -165,8 +165,8 @@ fn using_names_of_types_in_namespaces_works() {
 #[test]
 fn field_access_works_on_flipped_ports() {
     let code = r#"
-        struct port P {p1: &bool, p2: ~& bool}
-        entity t(p: ~P) -> bool {
+        struct port P {p1: &bool, p2: inv &bool}
+        entity t(p: inv P) -> bool {
             set p.p1 = true;
             *p.p2
         }