+++ CHANGELOG.md
@@ -19,7 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - [!346][!346] *Breaking change* Removed the `$comptime` system
-- [!350][!350] *Breaking change* Replaced `&mut` with `~&`
+- [!350][!350] *Breaking change* Replaced `&mut` with `inv &`
+- [!350][!350] *Breaking change* Changed the syntax of inverted ports from `~T` to `inv T`
 
 ### Fixed